+++ hari_client/client/client.py
@@ -1,10 +1,13 @@
 import datetime
 import json
+import os
 import pathlib
 import types
 import typing
 import uuid
 import warnings
+from concurrent import futures
+from urllib.parse import urlparse
 
 import pydantic
 import requests
@@ -575,7 +578,6 @@ class HARIClient:
         skip: int | None = None,
         query: models.QueryList | None = None,
         sort: list[models.SortingParameter] | None = None,
-        name_filter: str | None = None,
         archived: bool | None = False,
     ) -> list[models.DatasetResponse]:
         """Returns datasets that a user has access to.
@@ -587,7 +589,6 @@ class HARIClient:
             skip: skip the number of datasets returned
             query: query parameters to filter the datasets
             sort: sorting parameters to sort the datasets
-            name_filter: filter by dataset name
             archived: if true, return only archived datasets; if false (default), return non-archived datasets.
 
         Returns:
@@ -607,7 +608,6 @@ class HARIClient:
         self,
         visibility_statuses: tuple | None = (models.VisibilityStatus.VISIBLE,),
         query: models.QueryList | None = None,
-        name_filter: str | None = None,
         archived: bool | None = False,
     ) -> int:
         """
@@ -615,7 +615,6 @@ class HARIClient:
         Args:
             visibility_statuses: Visibility statuses of the returned datasets
             query: query parameters to filter the datasets
-            name_filter: filter by dataset name
             archived: if true, count only archived datasets; if false (default), count non-archived datasets.
 
         Returns:
@@ -1131,6 +1130,76 @@ class HARIClient:
         log.info(f"Fetched {len(medias)} medias successfully.")
 
         return medias
+
+    def download_media_files(
+        self,
+        dataset_id: uuid.UUID,
+        query: models.QueryList | None = None,
+        output_dir: str | None = os.getcwd(),
+        max_workers: int = 32,
+    ) -> None:
+        """Download media files of a dataset.
+
+        Args:
+            dataset_id: The dataset id
+            query: The filters to be applied to the search
+            output_dir: The directory where the media files will be downloaded (defaults to the current directory)
+            max_workers: The maximum number of threads to use for downloading media files (defaults to 32).
+
+        Raises:
+            APIException: If get medias request fails.
+        """
+        try:
+            medias = self.get_medias_paginated(
+                dataset_id=dataset_id,
+                batch_size=100,
+                query=query,
+            )
+            media_urls = [media.media_url for media in medias if media.media_url]
+        except errors.APIError as e:
+            log.error(f"Failed to fetch medias to download: {e}")
+            raise e
+
+        def download_single_media(url: str, output_dir: str):
+            name = urlparse(url).path.strip("/")
+            output_path = pathlib.Path(output_dir) / name
+            output_path.parent.mkdir(parents=True, exist_ok=True)
+
+            if output_path.exists():
+                return None
+
+            try:
+                with requests.get(url, stream=True) as response:
+                    response.raise_for_status()
+                    with open(output_path, "wb") as f:
+                        for chunk in response.iter_content(chunk_size=8192):
+                            if chunk:
+                                f.write(chunk)
+                return None
+            except requests.RequestException as e:
+                return url, str(e)
+
+        log.info("Starting to download media files...")
+        failed = []
+        with futures.ThreadPoolExecutor(max_workers=max_workers) as executor:
+            failed_url = {
+                executor.submit(download_single_media, url, output_dir): url
+                for url in media_urls
+            }
+
+            for future in tqdm(
+                futures.as_completed(failed_url),
+                total=len(media_urls),
+                desc="Downloading media files",
+            ):
+                result = future.result()
+                if result is not None:
+                    failed.append(result)
+
+        if failed:
+            log.warning(f"\nFailed to download {len(failed)} files:")
+            for url, error in failed:
+                log.warning(f"`{url}`: {error}")
 
     def archive_media(self, dataset_id: uuid.UUID, media_id: str) -> str:
         """Archive the media
@@ -2435,11 +2504,23 @@ class HARIClient:
 
     def get_multiple_aint_learning_data(
         self,
+        limit: int | None = None,
+        skip: int | None = None,
+        query: models.QueryList | None = None,
+        sort: list[models.SortingParameter] | None = None,
+        archived: bool | None = False,
     ) -> list[models.AINTLearningData]:
         """
         !!! Only available for qm internal users !!!
 
         Retrieve all AINT learning data available to the user.
+
+        Args:
+            limit: limit the number of AINT learning data returned
+            skip: skip the number of AINT learning data returned
+            query: query parameters to filter the AINT learning data
+            sort: sorting parameters to sort the AINT learning data
+            archived: whether to include archived AINT learning data
 
         Returns:
             A list of AINT learning data objects.
@@ -2447,6 +2528,7 @@ class HARIClient:
         return self._request(
             "GET",
             f"/aintLearningData",
+            params=self._pack(locals()),
             success_response_item_model=list[models.AINTLearningData],
         )
 
@@ -2580,16 +2662,50 @@ class HARIClient:
             success_response_item_model=str,
         )
 
+    def get_aint_learning_data_count(
+        self,
+        query: models.QueryList | None = None,
+        archived: bool | None = False,
+    ) -> int:
+        """
+        Returns aint learning data count for the user.
+        Args:
+            query: query parameters to filter the aint learning datas
+            archived: if true, count only archived aint learning data; if false (default), count non-archived aint learning data.
+
+        Returns:
+            The number of aint learning data
+
+        Raises:
+            APIException: If the request fails.
+        """
+        return self._request(
+            "GET",
+            "/aintLearningData:count",
+            params=self._pack(locals()),
+            success_response_item_model=int,
+        )
+
     def get_ml_annotation_models(
         self,
         projection: dict[str, bool] | None = None,
+        limit: int | None = None,
+        skip: int | None = None,
+        query: models.QueryList | None = None,
+        sort: list[models.SortingParameter] | None = None,
+        archived: bool | None = False,
     ) -> list[models.MlAnnotationModel]:
         """
-        Retrieve all ml annotation models available to the user.
+        Retrieve ml annotation models available to the user.
 
         Args:
             projection: The fields to be returned (dictionary keys with value True are returned,
             keys with value False are not returned).
+            limit: limit the number of ml annotation models returned
+            skip: skip the number of ml annotation models returned
+            query: query parameters to filter the ml annotation models
+            sort: sorting parameters to sort the ml annotation models
+            archived: whether to include archived ml annotation models
 
         Returns:
              A list of ml annotation models.
@@ -2768,11 +2884,47 @@ class HARIClient:
             success_response_item_model=str,
         )
 
+    def get_ml_annotation_model_count(
+        self,
+        query: models.QueryList | None = None,
+        archived: bool | None = False,
+    ) -> int:
+        """
+        Returns ml annotation model count for the user.
+        Args:
+            query: query parameters to filter the ml annotation models
+            archived: if true, count only archived ml annotation models; if false (default), count non-archived ml annotation models.
+
+        Returns:
+            The number of ml annotation models
+
+        Raises:
+            APIException: If the request fails.
+        """
+        return self._request(
+            "GET",
+            "/mlAnnotationModels:count",
+            params=self._pack(locals()),
+            success_response_item_model=int,
+        )
+
     def get_ai_annotation_runs(
         self,
+        limit: int | None = None,
+        skip: int | None = None,
+        query: models.QueryList | None = None,
+        sort: list[models.SortingParameter] | None = None,
+        archived: bool | None = False,
     ) -> list[models.AIAnnotationRun]:
         """
-        Retrieve all AI annotation runs available to the user.
+        Retrieve AI annotation runs available to the user.
+
+        Args:
+            limit: limit the number of AI annotation runs returned
+            skip: skip the number of AI annotation runs returned
+            query: query parameters to filter the AI annotation runs
+            sort: sorting parameters to sort the AI annotation runs
+            archived: whether to include archived AI annotation runs
 
         Returns:
             A list of AI annotation runs.
@@ -2780,6 +2932,7 @@ class HARIClient:
         return self._request(
             "GET",
             f"/aiAnnotationRuns",
+            params=self._pack(locals()),
             success_response_item_model=list[models.AIAnnotationRun],
         )
 
@@ -2911,11 +3064,49 @@ class HARIClient:
             success_response_item_model=str,
         )
 
+    def get_ai_annotation_run_count(
+        self,
+        query: models.QueryList | None = None,
+        archived: bool | None = False,
+    ) -> int:
+        """
+        Returns ai annotation run count for the user.
+        Args:
+            query: query parameters to filter the ai annotation runs
+            archived: if true, count only archived ai annotation runs; if false (default), count non-archived ai annotation runs.
+
+        Returns:
+            The number of ai annotation runs
+
+        Raises:
+            APIException: If the request fails.
+        """
+        return self._request(
+            "GET",
+            "/aiAnnotationRuns:count",
+            params=self._pack(locals()),
+            success_response_item_model=int,
+        )
+
     ### pipelines ###
 
-    def get_pipelines(self) -> list[models.Pipeline]:
+    def get_pipelines(
+        self,
+        limit: int | None = None,
+        skip: int | None = None,
+        query: models.QueryList | None = None,
+        sort: list[models.SortingParameter] | None = None,
+        archived: bool | None = False,
+    ) -> list[models.Pipeline]:
         """
-        Get all pipelines.
+        Get pipelines available for the user.
+
+        Args:
+            limit: limit the number of pipelines returned
+            skip: skip the number of pipelines returned
+            query: query parameters to filter the pipelines
+            sort: sorting parameters to sort the pipelines
+            archived: whether to include archived pipelines
 
         Returns:
             A list of pipeline objects.
@@ -2923,6 +3114,7 @@ class HARIClient:
         return self._request(
             "GET",
             "/pipelines",
+            params=self._pack(locals()),
             success_response_item_model=list[models.Pipeline],
         )
 
@@ -2939,11 +3131,49 @@ class HARIClient:
             success_response_item_model=models.PipelineWithNodes,
         )
 
+    def get_pipeline_count(
+        self,
+        query: models.QueryList | None = None,
+        archived: bool | None = False,
+    ) -> int:
+        """
+        Returns pipeline count for the user.
+        Args:
+            query: query parameters to filter the pipelines
+            archived: if true, count only archived pipelines; if false (default), count non-archived pipelines.
+
+        Returns:
+            The number of pipelines
+
+        Raises:
+            APIException: If the request fails.
+        """
+        return self._request(
+            "GET",
+            "/pipelines:count",
+            params=self._pack(locals()),
+            success_response_item_model=int,
+        )
+
     ### annotation runs ###
 
-    def get_annotation_runs(self) -> list[models.AnnotationRun]:
+    def get_annotation_runs(
+        self,
+        limit: int | None = None,
+        skip: int | None = None,
+        query: models.QueryList | None = None,
+        sort: list[models.SortingParameter] | None = None,
+        archived: bool | None = False,
+    ) -> list[models.AnnotationRun]:
         """
-        Get all annotation runs.
+        Get annotation runs available to the users.
+
+        Args:
+            limit: limit the number of annotation runs returned
+            skip: skip the number of annotation runs returned
+            query: query parameters to filter the annotation runs
+            sort: sorting parameters to sort the annotation runs
+            archived: whether to include archived annotation runs
 
         Returns:
             A list of annotation run objects.
@@ -2951,6 +3181,7 @@ class HARIClient:
         return self._request(
             "GET",
             "/annotationRuns",
+            params=self._pack(locals()),
             success_response_item_model=list[models.AnnotationRun],
         )
 
@@ -2986,4 +3217,28 @@ class HARIClient:
             "/annotationRuns",
             json=annotation_run.model_dump(),
             success_response_item_model=models.AnnotationRun,
+        )
+
+    def get_annotation_run_count(
+        self,
+        query: models.QueryList | None = None,
+        archived: bool | None = False,
+    ) -> int:
+        """
+        Returns annotation run count for the user.
+        Args:
+            query: query parameters to filter the annotation runs
+            archived: if true, count only archived annotation runs; if false (default), count non-archived annotation runs.
+
+        Returns:
+            The number of annotation runs
+
+        Raises:
+            APIException: If the request fails.
+        """
+        return self._request(
+            "GET",
+            "/annotationRuns:count",
+            params=self._pack(locals()),
+            success_response_item_model=int,
         )