+++ CHANGELOG.md
@@ -1,7 +1,29 @@
 ## [major.minor.patch] - DD-MM-YYYY
 
+## Breaking Changes
+
+- Deleted `name_filter` from get multiple datasets endpoint [PR#91](https://github.com/quality-match/hari-client/pull/91)
+
 ### New Features
 
+- Add pagination, sorting and filtering to the get annotation runs, pipelines, AINT learning data, ML annotation models, AI annotation runs methods [PR#91](https://github.com/quality-match/hari-client/pull/91)
+- Add count annotation runs, pipelines, AINT learning data, ML annotation models, AI annotation runs methods [PR#91](https://github.com/quality-match/hari-client/pull/91)
+- Add `ilike` query operator [PR#91](https://github.com/quality-match/hari-client/pull/91)
+
+- Add method to download media files from dataset [PR#94](https://github.com/quality-match/hari-client/pull/94)
+
+#### Consistency in partially failed uploads
+
+- partially failed bulk uploads of medias and media objects are now handled consistently [PR#80](https://github.com/quality-match/hari-client/pull/80)
+  - for failed media uploads, any media_objects and attributes for these medias will not be tried to be uploaded
+  - for media_objects, any attributes for these media_objects will not be tried to be uploaded
+
+#### Better reporting of failed and skipped uploads
+
+- improved reporting of failed and skipped uploads [PR#80](https://github.com/quality-match/hari-client/pull/80)
+  - failed uploads are now reported in the `HARIUploadResults.failures` field
+
+#### 3D
 - added automatic scene creation to HARIUploader to support setting up a dataset with 3D data [PR#86](https://github.com/quality-match/hari-client/pull/86)
   - added `quickstart_3D.py` example script to show how to upload 3D data with the HARIUploader
   - disclaimer: there's no example data for this script yet, so it won't work out of the box.